refactor: consolidate split layout modules
Rename the 3 split layout files for clearer roles:
- split-helpers.js → split-primitives.js (pure geometry functions)
- split-layout-ops.js → split-layout.js (DOM tree manipulation)
- terminal-split-ops.js → terminal-split.js (terminal-specific operations)

Update all imports across the project accordingly.

Closes #95
