+++ src/utils/split-primitives.js
